# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestão de Estacionamentos
  description: API para gerenciamento de vagas, entrada/saída de veículos e faturamento
  version: 1.0.0
servers:
  - url: http://localhost:3003
    description: Servidor local
paths:
...
  /spot-status:
    post:
      tags:
        - Query
      summary: Consulta status de uma vaga
      operationId: getSpotStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotStatusRequest'
      responses:
        '200':
          description: Status da vaga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotStatusResponse'
        '404':
          description: Vaga não encontrada
...
components:
  schemas:
...
    SpotStatusRequest:
      type: object
      required:
        - lat
        - lng
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
    SpotStatusResponse:
      type: object
      properties:
        ocupied:
          type: boolean
        license_plate:
          type: string
        price_until_now:
          type: number
          format: double
        entry_time:
          type: string
          format: date-time
        time_parked:
          type: string
          format: date-time
...